+++ src/main/java/com/global/api/network/enums/DE48_CustomerDataType.java
@@ -4,14 +4,13 @@ import com.global.api.entities.enums.IStringConstant;
 
 public enum DE48_CustomerDataType implements IStringConstant {
     UnencryptedIdNumber("0"),
-    Vehicle_Trailer_Number("1"),
+    Vehicle_Number("1"),
     VehicleTag("2"),
     DriverId_EmployeeNumber("3"),
-    Odometer_Hub_Reading("4"),
+    Odometer_Reading("4"),
     DriverLicense_Number("5"),
     DriverLicense_State_Province("6"),
     DriverLicense_Name("7"),
-    WorkOrder_PoNumber("8"),
     InvoiceNumber("9"),
     TripNumber("A"),
     UnitNumber("B"),
@@ -27,7 +26,10 @@ public enum DE48_CustomerDataType implements IStringConstant {
     JobNumber("U"),
     Department("V"),
     LoyaltyInformation("W"),
-    MerchantOrderNumber("Z");
+    MerchantOrderNumber("Z"),
+    MaintenanceNumber("a"),
+    TrailerNumber("b"),
+    HubometerNumber("c");
 
     private final String value;
     DE48_CustomerDataType(String value) { this.value = value; }