+++ src/test/java/com/global/api/tests/network/vaps/VapsTimeRequestTests.java
@@ -1,0 +1,72 @@
+package com.global.api.tests.network.vaps;
+
+import com.global.api.ServicesContainer;
+import com.global.api.entities.Transaction;
+import com.global.api.entities.exceptions.ApiException;
+import com.global.api.network.enums.*;
+import com.global.api.serviceConfigs.AcceptorConfig;
+import com.global.api.serviceConfigs.NetworkGatewayConfig;
+import com.global.api.services.NetworkService;
+import com.global.api.tests.BatchProvider;
+import com.global.api.tests.StanGenerator;
+import org.junit.FixMethodOrder;
+import org.junit.Test;
+import org.junit.runners.MethodSorters;
+
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class VapsTimeRequestTests {
+
+    public VapsTimeRequestTests() throws ApiException {
+
+        AcceptorConfig acceptorConfig = new AcceptorConfig();
+
+        // data code values
+        acceptorConfig.setCardDataInputCapability(CardDataInputCapability.ContactlessEmv_ContactEmv_ContactlessMsd_MagStripe_KeyEntry);
+        acceptorConfig.setCardHolderAuthenticationCapability(CardHolderAuthenticationCapability.PIN);
+        acceptorConfig.setCardHolderAuthenticationEntity(CardHolderAuthenticationEntity.ByMerchant);
+        acceptorConfig.setTerminalOutputCapability(TerminalOutputCapability.Printing_Display);
+
+        // hardware software config values
+        acceptorConfig.setHardwareLevel("34");
+        acceptorConfig.setSoftwareLevel("21205710");
+
+        // pos configuration values
+        acceptorConfig.setSupportsPartialApproval(true);
+        acceptorConfig.setSupportsShutOffAmount(true);
+        acceptorConfig.setSupportsReturnBalance(true);
+        acceptorConfig.setSupportsDiscoverNetworkReferenceId(true);
+        acceptorConfig.setSupportsAvsCnvVoidReferrals(true);
+        acceptorConfig.setSupportsEmvPin(true);
+
+        // gateway config
+        NetworkGatewayConfig config = new NetworkGatewayConfig();
+        config.setPrimaryEndpoint("test.txns-c.secureexchange.net");
+        config.setPrimaryPort(15031);
+        config.setSecondaryEndpoint("test.txns.secureexchange.net");
+        config.setSecondaryPort(15031);
+        config.setCompanyId("0044");
+        config.setTerminalId("0000912197711");
+        config.setAcceptorConfig(acceptorConfig);
+        config.setEnableLogging(true);
+        config.setStanProvider(StanGenerator.getInstance());
+        config.setBatchProvider(BatchProvider.getInstance());
+
+        ServicesContainer.configureService(config);
+
+        // with merchant type
+        config.setMerchantType("5542");
+        ServicesContainer.configureService(config, "ICR");
+    }
+
+    @Test
+    public void test_000_TimeRequest() throws ApiException {
+        Transaction response = NetworkService.timeRequest()
+                .execute();
+        assertNotNull(response.getIssuerData().get(CardIssuerEntryTag.HeartlandTimeRequest));
+    }
+
+
+}